Relocate commonFunctions
