+++ __tests__/fixtures/commonFunctions.ts
